+++ daemons/utils.py
@@ -165,3 +165,17 @@ def load_json_dict(s, error_message):
         except json.JSONDecodeError as e:
             raise Exception(error_message) from e
     return json_dict
+
+
+def is_int(v):
+    try:
+        int(v)
+        return True
+    except Exception:
+        return False
+
+
+def try_cast_num(v):
+    if is_int(v):
+        return int(v)
+    return v